moved validator into new dir and make root readme
