+++ DIFI_Validator/docker/input/start.sh
