+++ DIFI_Validator/python/dificommon.py
